+++ src/main/java/woowacourse/shoppingcart/dto/request/UpdateCartRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.request;
 
 import woowacourse.shoppingcart.domain.Cart;
