+++ src/main/java/woowacourse/shoppingcart/dto/request/UpdateCartRequests.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.request;
 
 import java.util.List;
 import java.util.stream.Collectors;
