+++ src/main/java/woowacourse/shoppingcart/dto/response/SignUpResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.response;
 
 import woowacourse.shoppingcart.domain.Customer;
